Playing around with the idea of option contexts
This currently does not run due to a circular import error. That
will be easy enough to work around, but after looking at this
more I don't think the added complexity will be worth it. I do
not believe that many people would use these features at all, and
it would make the code, documentation, unit testing, etc all more
complex. For now I am going to commit and sholder this, but keep
the code around in case I decide it is worth it to implement this
in the future.

What prompted this was the an issue asking for seperate secrets
on seperate endpoints, a totally reasonable request. For now I am
going to implement that via a kwarg passed into the view decorators,
which will keep the code simplier. If more features like this get
requested, i will consider adding more kwargs, or looking at the
option contexts again.